Fixed tests. Added few mehtods to social_network. #796
